requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.6
certifi==2020.6.20
chardet==3.0.4
cos-python-sdk-v5==1.8.0
dicttoxml==1.7.4
Django==3.0.7
django-cors-headers==3.4.0
django-filter==2.3.0
django-redis==4.12.1
djangorestframework==3.11.0
future==0.18.2
idna==2.9
importlib-metadata==1.7.0
kombu==4.6.11
more-itertools==8.4.0
python-crontab==2.5.1
python-dateutil==2.8.1
pytz==2020.1
qcloud-python-sts==3.0.3
redis==3.5.3
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
tencentcloud-sdk-python==3.0.203
urllib3==1.25.9
vine==1.3.0
XML2Dict==0.2.2
zipp==3.1.0
